مسابقات/در پیچ و خم پایتون/کارآگاه آکا/requirements.txt
django
djangorestframework
-r python_requirements.txt
